convert_mesh.py: new --extract-edges option in main()
- move merge_lines(), extract_edges() into mesh_tools.py
- remove extract_edges.py
